# Backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/backorders": {
      "get": {
        "operationId": "BackordersController_listBackorders_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limit",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "SortBy",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "SortOrder",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "StartDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "EndDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Status",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Backorder"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List Backorders",
        "tags": [
          "Backorders"
        ]
      }
    },
    "/v1/backorders/{id}": {
      "get": {
        "operationId": "BackordersController_getBackorder_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Backorder",
        "tags": [
          "Backorders"
        ]
      }
    },
...
    "/v1/orders": {
      "post": {
        "operationId": "OrdersController_createOrder_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create Order (SALESMAN only)",
        "tags": [
          "Orders"
        ]
      },
      "get": {
        "operationId": "OrdersController_getOrders_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limit",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "SortBy",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "SortOrder",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "StartDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "EndDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Status",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "salesman_id",
            "required": false,
            "in": "query",
            "description": "Filter by salesman UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shop_id",
            "required": false,
            "in": "query",
            "description": "Filter by shop UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Order"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Orders (paginated, filtered, searched)",
        "tags": [
          "Orders"
        ]
      }
    },
    "/v1/orders/distributor-to-manufacturer": {
      "post": {
        "description": "Accepts a single payload of products and splits them into separate orders based on the manufacturer. Distributor self-products (manufacturer=null) are grouped into their own order. The provided bill discount is applied independently to each created order. Returns an array of created orders.",
        "operationId": "OrdersController_createDistributorManufacturerOrder_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDistributorManufacturerOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create Distributor Order (DISTRIBUTOR_ADMIN only)",
        "tags": [
          "Orders"
        ]
      }
    },
    "/v1/orders/purchase-request/generate": {
      "post": {
        "operationId": "OrdersController_generatePurchaseRequest_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate Purchase Request (DISTRIBUTOR_ADMIN only)",
        "tags": [
          "Orders"
        ]
      }
    },
    "/v1/orders/export/csv": {
      "get": {
        "operationId": "OrdersController_exportCsv_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limit",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "SortBy",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "SortOrder",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "StartDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "EndDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Status",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "salesman_id",
            "required": false,
            "in": "query",
            "description": "Filter by salesman UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shop_id",
            "required": false,
            "in": "query",
            "description": "Filter by shop UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Export Orders as CSV",
        "tags": [
          "Orders"
        ]
      }
    },
    "/v1/orders/export/xlsx": {
      "get": {
        "operationId": "OrdersController_exportXlsx_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limit",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "SortBy",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "SortOrder",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "StartDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "EndDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Status",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "salesman_id",
            "required": false,
            "in": "query",
            "description": "Filter by salesman UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shop_id",
            "required": false,
            "in": "query",
            "description": "Filter by shop UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Export Orders as Excel (XLSX)",
        "tags": [
          "Orders"
        ]
      }
    },
    "/v1/orders/backorders": {
      "get": {
        "operationId": "OrdersController_getBackorders_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limit",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "SortBy",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "SortOrder",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "StartDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "EndDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Status",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "distributor_id",
            "required": false,
            "in": "query",
            "description": "Filter by specific distributor UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "salesman_id",
            "required": false,
            "in": "query",
            "description": "Filter by specific salesman UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Backorders (paginated)",
        "tags": [
          "Orders"
        ]
      }
    },
    "/v1/orders/backorders/{id}": {
      "get": {
        "operationId": "OrdersController_getBackorderById_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Backorder by ID",
        "tags": [
          "Orders"
        ]
      }
    },
    "/v1/orders/backorders/{id}/resolve": {
      "patch": {
        "operationId": "OrdersController_resolveBackorder_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveBackorderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Resolve Backorder",
        "tags": [
          "Orders"
        ]
      }
    },
    "/v1/orders/{id}": {
      "get": {
        "operationId": "OrdersController_getOrderById_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Order By Id (includes line items)",
        "tags": [
          "Orders"
        ]
      },
      "patch": {
        "operationId": "OrdersController_updateOrder_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit Order (SALESMAN/DISTRIBUTOR_ADMIN/MANUFACTURER_ADMIN)",
        "tags": [
          "Orders"
        ]
      }
    },
    "/v1/orders/{id}/cancel": {
      "patch": {
        "operationId": "OrdersController_cancelOrder_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel Order",
        "tags": [
          "Orders"
        ]
      }
    },
    "/v1/orders/{id}/status": {
      "patch": {
        "description": "CREATED → CONFIRMED → PROCESSING → PACKED → DISPATCHED → DELIVERED. CANCELLED is via the /cancel endpoint.",
        "operationId": "OrdersController_updateOrderStatus_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update Order Status (lifecycle transition)",
        "tags": [
          "Orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vikas Inventory RN Platform API",
    "description": "Enterprise Backend API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginatedDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginatedMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "Backorder": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Id"
          },
          "order_id": {
            "type": "string",
            "description": "Order id"
          },
          "order_item_id": {
            "type": "string",
            "description": "Order item id"
          },
          "product_id": {
            "type": "string",
            "description": "Product id"
          },
          "distributor_id": {
            "type": "string",
            "description": "Distributor id"
          },
          "quantity": {
            "type": "number",
            "description": "Quantity"
          },
          "status": {
            "type": "string",
            "description": "Status"
          },
          "resolved_quantity": {
            "type": "number",
            "description": "Resolved quantity"
          },
          "resolved_at": {
            "format": "date-time",
            "type": "string",
            "description": "Resolved at"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Created at"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "description": "Updated at"
          }
        },
        "required": [
          "id",
          "order_id",
          "product_id",
          "distributor_id",
          "quantity",
          "status",
          "resolved_quantity",
          "created_at",
          "updated_at"
        ]
      },
...
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "visitId": {
            "type": "string",
            "description": "VisitId"
          },
          "shopId": {
            "type": "string",
            "description": "ShopId"
          },
          "products": {
            "description": "Products",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isOfflineCreated": {
            "type": "boolean",
            "description": "IsOfflineCreated"
          },
          "idempotencyKey": {
            "type": "string",
            "description": "IdempotencyKey"
          },
          "transportMode": {
            "type": "string",
            "description": "Transport Mode"
          },
          "standardDiscountPercent": {
            "type": "number",
            "description": "Standard Discount Percent"
          },
          "specialDiscountPercent": {
            "type": "number",
            "description": "Special Discount Percent"
          }
        },
        "required": [
          "visitId",
          "shopId",
          "products"
        ]
      },
      "CreateDistributorManufacturerOrderDto": {
        "type": "object",
        "properties": {
          "products": {
            "description": "Products",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "idempotencyKey": {
            "type": "string",
            "description": "IdempotencyKey"
          },
          "transportMode": {
            "type": "string",
            "description": "Transport Mode"
          }
        },
        "required": [
          "products"
        ]
      },
...
      "ResolveBackorderDto": {
        "type": "object",
        "properties": {
          "resolved_quantity": {
            "type": "number",
            "description": "The quantity being resolved"
          },
          "notes": {
            "type": "string",
            "description": "Optional notes regarding the resolution"
          }
        },
        "required": [
          "resolved_quantity"
        ]
      },
...
    }
  }
}